Isolate inputs/outputs as separate packages
