+++ internal/output/kubernetesTokenReview/output.go
@@ -1,4 +1,4 @@
-package output
+package kubernetesTokenReview
 
 import (
 	"encoding/json"
@@ -10,12 +10,12 @@ import (
 	"github.com/robbiemcmichael/auth-mux/internal/types"
 )
 
-type KubernetesTokenReview struct {
+type Output struct {
 	Audience string `yaml:"audience"`
 	MaxTTL   int64  `yaml:"maxTTL"`
 }
 
-func (o *KubernetesTokenReview) Handler(w http.ResponseWriter, validation types.Validation) error {
+func (o *Output) Handler(w http.ResponseWriter, validation types.Validation) error {
 	tokenReview := auth.TokenReview{
 		TypeMeta: metav1.TypeMeta{
 			APIVersion: auth.SchemeGroupVersion.String(),
